int i in for boucle => size_t

+++ Projet/AsianCallOption.cpp
@@ -11,7 +11,7 @@ OptionType AsianCallOption::GetOptionType() const { return OptionType::Call; }
 double AsianCallOption::payoffPath(vector<double> past_prices) const {
 	cout << "payoffPath in";
 	double price = 0;
-	for (int i = 0; i < past_prices.size(); i++) {
+	for (size_t i = 0; i < past_prices.size(); i++) {
 		price += past_prices[i];
 	}
